requirementst.txt
loguru==0.5.3
flask==2.0.2
flake8==4.0.1
black==21.9b0
isort==5.9.3
peewee==3.14.4
pymysql==1.0.2
cryptography==35.0.0
flask-jwt-extended==4.3.1
flask-cors==3.0.10
redis==4.0.2
ruamel.yaml==0.17.16
flask-cors==3.0.10
elastic-transport==8.0.1
elasticsearch==8.0.0
celery==5.2.3
minio==7.1.4
numpy==1.18.5
python-redis-lock==3.7.0
requests==2.25.1
Flask-APScheduler==1.12.3
flask-sqlalchemy==2.5.1
eventlet==0.30.2